Always generate perfbook.synctex.gz
Commit 8458f7572335 ("Add optional flag 'LATEX_OPT' in run{first}latex.sh")
was ment to support optinal generation of .synctex files.
It turns out that evince and gedit pair supports both the forward
(.tex -> PDF) and the inverse (PDF -> .tex) searches [1].
I find this mechanism helps a lot in perfbook, because of its
nested LaTeX source structure.

For forward search (.tex -> PDF) to work, each sub-sources needs
to have a pointer to main file in the first/last 3 lines. This
commit therefor adds the pointers to top level perfbook.tex.

Also add an entry in FAQ-BUILD.txt as an introduction of the
capability.

Note: As SyncTeX search will work only for perfbook.pdf, other
targets do not generate .synctex.gz.

[1]: https://help.gnome.org/users/evince/stable/synctex-search.html.en

Signed-off-by: Akira Yokosawa <akiyks@gmail.com>