Reorganize backmatters
Move Glossary, Bibliography, Credits, and Index chapters to backmatter.
Chapter and section counting is disabled there.

This change accomodates the fact that LaTeX packages which support
Bibliograpy and Index works well if they are placed in backmatter area.

By the help of "tocbibind" package, we can reference those chapters
from TOC.

NOTE: Moving backmatter contents in appendix/ to perfbook-lt.tex
or a top-level .tex file is desiable in the near future.

Signed-off-by: Akira Yokosawa <akiyks@gmail.com>