PoC of indexing
Add Index at the end of the book for the new target "perfbook-ix.pdf".
Simple index markers are put in glossary.tex for testing the changes
in the build framework.

Note that we need to run the command "makeindex" at every pdflatex
iteration due to the fact that it extracts page numbers corresponding
to .aux file at the time it runs rather than references to pages.
As page numbers can change in later iterations, it is added in
runlatex.sh's "iterate_latex ()".

Signed-off-by: Akira Yokosawa <akiyks@gmail.com>