Reorganize backmatters
Move Glossary, Bibliography, Credits, and Index chapters to back matter.
Chapter and section counters are disabled there.

This change accomodates the fact that LaTeX packages supporting
Bibliograpy and Index work well if they are placed in the back-matter
area.

The tocbibind package is one of such packages and allows us to reference
back-matter chapters from TOC.

Signed-off-by: Akira Yokosawa <akiyks@gmail.com>