PoC of addtional API Index
By using "splitidx" package instead of "makeidx", one can generate
multiple index raw files.
Employ this facility to make an Index of API.

To indicate the category of each API, each annotation of API in the
.tex files is done by one of new commands defined in this commit,
\apic (for C11), \apig (for GCC), \apipx (for POSIX), \apipf (for
perfbook), etc.

Commands \apialtc, \apialtg, and \apialtk are for cases where original
function in \co{} macro has some arguments.
You can give a function API without arguments in the 2nd argument.

Appearence of API category in the Index page can be improved by
tweaking the definition of the macro \categapi in the preamble.

Signed-off-by: Akira Yokosawa <akiyks@gmail.com>