index: Add annotations to people's names for PoC
Not complete.
Names are included in the general Index pages for the moment.

In the Index, names are presented in "surname, forename" style,
as the same way as in .bib file's author fields.

Also, to help adding annotations for indexing, add a new boolean
"indexhl" and define \hlindex{} command in \IfIndexHl{}{}.

The definintion is to be tweaked in a later change.

Signed-off-by: Akira Yokosawa <akiyks@gmail.com>