Use 2c-layout for acronyms in both 1c- and 2c- layout builds
For ebook-size build, use 1c-layout instead.

While old code modified textwidth and column number for 1c builds,
new code uses common textwidth and 2c- layout for acronym list
except for the ebook-size build.

There is some room of improvemnt in the appearence of acronym list.
It would be desirable for the list to have balanced looks similar
to that of index pages.

Signed-off-by: Akira Yokosawa <akiyks@gmail.com>