cleverefcheck: Add check of lower-case word starting sentence
As cleverefcheck knows of "new_sentence", flag lower-case word
at the beginning of an input line.

Footnote is a special case.

Try to recognize mid-sentence footnote by $in_footnote and
$footnote_save.

Colons can have both lower-case and upper-case words to follow.
So we don't care the next word of a colon is lower-case.

Signed-off-by: Akira Yokosawa <akiyks@gmail.com>