EXP Makefile: Don't gobble messages from inkscape
This is for debugging purpose and not intented to be merged.

Yes, silently gobbling warning/error messages from inkscape is not
a good idea in general, but for building perfbook, we don't need
full set of inkscape and such inkscape outputs harmless warnings.

So, in normal cases, gobbling such messages looks a reasonable
thing to do for us.

Signed-off-by: Akira Yokosawa <akiyks@gmail.com>