run{first}latex.sh: Use variable for 'pdflatex' command
In testing a dev version of upstream LaTeX, "pdflatex-dev" is used
in place of "pdflatex".
Hardcoded "pdflatex"s in run{first}latex.sh need editing in such
testing.

Using a variable named "LATEX", whose default is pdflatex, in the
scripts helps in those cases.

Update Makefile as well so that the dev version can be specified
as follows:

    make LATEX=pdflatex-dev

Note 1: The trick for setting a default value to a shell variable
using ":" is borrowed from an answer on unix.stackexchange.com:

    https://unix.stackexchange.com/questions/122845/

Note 2: For regression test of utilities/parallel-latex-regress.sh,
make copies of scripts before this change as
utilities/runfirstlatex.2022.01.25.sh and utilities/runlatex.2022.01.25.

Signed-off-by: Akira Yokosawa <akiyks@gmail.com>