run{first}latex.sh: Use variable for 'pdflatex' command
In testing a dev version of upstream LaTeX, "pdflatex-dev" is used
in place of "pdflatex".
Hardcoded "pdflatex"s in run{first}latex.sh need editing in such
testing.

Using a variable "LATEXCMD", whose default is pdflatex, in the
scripts helps in those cases.

Update Makefile as well so that the dev version can be specified
as follows:

    make LATEX=pdflatex-dev

Signed-off-by: Akira Yokosawa <akiyks@gmail.com>