Makefile: Add pattern for newtxtext < 1.7
newtxtext < 1.7 doesn't have "nofontspec" option.
It behaves the same as newtxtext >= 1.7 with "nofontspec" on LuaLaTeX.
Check the version and remove nofontspec on the fly to prevent
build errors.

As the on-the-fly change involves perfbook-lt.tex itself,
rename it perfbook-main.tex.  Now, perfbook-lt.tex is generated
when necessary.

Recipe of target paper-clean is changed so that it cleans
only perfbook-*.tex files directly generated from perfbook-main.tex.

Signed-off-by: Akira Yokosawa <akiyks@gmail.com>