Makefile: Add pattern for newtxtext < 1.7
newtxtext >= 1.7 behaves differently from < 1.7 when the same
\usepackage[...]{newtxtext} command is used under LuaLaTeX
or XeLaTeX.

By adding the option of "nofontspec", it restores the existing font
choices.
Check the version and add nofontspec on the fly to this effect.

As the on-the-fly change involves perfbook-lt.tex itself,
rename it perfbook-main.tex.  Now, perfbook-lt.tex is generated
when necessary.

The recipe of target paper-clean is changed so that it cleans
only perfbook-*.tex files directly generated from perfbook-main.tex.

Also update parallel-latex-regress.sh so that perfbook-lt.tex is
present before invoking the reference Makefile.

Signed-off-by: Akira Yokosawa <akiyks@gmail.com>