fontconfig.sty: Partly restore even space ellipsis (unicode LaTeX)
By default, XeLaTeX and LuaLaTeX substitutes Unicode ELLIPSIS
code for \dots/\ldots.  Depending on the font design, this results
in uneven looking ellipsis.

This issue is mentioned in a blog post at:
    https://www.preppylion.com/the-pitfalls-of-unicode-choices-for-some-latex-symbols/

Borrow the workaround to prevent the substituion in text mode.

This has no effect in math mode, and ends up in uneven spacing there.

Signed-off-by: Akira Yokosawa <akiyks@gmail.com>