Use fig2dev as fallback of fig2eps
fig2ps as a distro package is not widely available on Linux distros.
Ubuntu/Debian and Fedora/RHEL/CentOS are exceptional in this regard.

Users of other distros need to download a source tar ball from
sourceforge.net (or fossies.org among others) and manually install
it.

fig2ps would be required if there were LaTeX special text in FIG
files. But that is not the case and fig2dev should suffice.

Add rules in Makefile so that fig2dev can be used for .fig --> .eps
conversion when fig2eps is not found.

Add scripts for fixing fonts as well. Only a dozen of substitution
patterns are needed for EPS from fig2dev with regard to existing
FIG files.

Note:
    fig2dev is often provided in transfig package. fig2ps requires
    fig2dev.

Signed-off-by: Akira Yokosawa <akiyks@gmail.com>