Fallback to fig2dev when fig2eps not found
fig2ps as a distro package is not widely available on Linux distros.
Ubuntu/Debian and Fedora/RHEL/CentOS are exceptional in this regards.

Users of other distros need to download source tar ball from
sourceforge.net or fossies.org among others and manually install
it.

For converting FIG into EPS, fig2dev (or transfig) is good enough.

Add rules in Makefile so that fig2dev can be used in .fig --> .eps
conversion instead when fig2eps is not found.

Add scripts for fixing fonts as well. Only a few substitution
patters are needed for EPS from fig2dev.

Signed-off-by: Akira Yokosawa <akiyks@gmail.com>