Fix warning of missing font shape
Since commit 6be27f7414d8 ("Use Biolinum font family for epigraph"),
new warning messages appear in the .log files as the following:

    LaTeX Font Warning: Font shape `OT1/minntx/sb/n' undefined
    (Font)              using `OT1/minntx/m/n' instead on input line 622.

    LaTeX Font Warning: Some font shapes were not available, defaults
    substituted.

This is because netwxmath package detects boilinum/libertine fonts
which has semi-bold shape and set the option of semi-bold shape
for the default font of newtxtext.

This could be a bug in the newtx package, but can be avoided by
moving the load of biolinum after the load of newtxmath package.

After loading biolinum, restore \rmdefault of newtxtext.

Signed-off-by: Akira Yokosawa <akiyks@gmail.com>