Use type1 font settings for all latex engines
Since v1.7, newtxtext has acquired OpenType font capability.
This breaks existing perfbook sources on LuaLaTeX and XeLaTeX
engines.

Our goal at the memoent is to make them buildable with LuaLaTeX
to suppress harmless warnings with regard to tcolorbox >=6.0.
Therefore, keep using type1 font settings on all latex engines.
newtxtext >=v1.7 has an option named "nofontspec" so that its type1
font setting for PDFLaTeX can be enabled.
biolinum package has an option "type1" for the same purpose.

One caveat in newtxtext is that nofontspec can not be specified
with newtxtext <1.7.
For the portablity, add code in autodate.sh to detect newtxtext >=1.7
and to emit LaTeX code setting an boolean, which determines the
\usepackage{newtxtext} command to be used.

As autodate.tex includes code assuming tcolorbox and the custom
command of \IfQqzBg, move related commands along with the
\input{autodate.tex} in front of \usepackage{newtxtext}.

Signed-off-by: Akira Yokosawa <akiyks@gmail.com>