Get rid of 'twocolumn' option in \documentclass
There is a lingering warning from the multicol package (indirectly
required by the idxlayout package) which reads:

    Package multicol Warning: May not work with the twocolumn
    option on input line 143.

This is due to the twocolumn option in \documentclass at the top
of preamble.

It turns out that it can be removed with minor tweaks.

One is the setting of \twocolumnwidth using \columnwidth for
2c builds.  Column width set by geometry becomes accessible after
\begin{document}.  Previously, the twocolumn option at
\documentclass permitted its use in the middle of preamble.

Instead, always set it using the value of 3.125in.

Another is the use of \sloppy.  It is implied in the book class
when the twocolumn option is specified.  Its effect is to prefer
wide inter-word white spaces to horizontal overfulls.
For perfbook, \sloppy is used explicitly for eb builds.
Do the same for 2c builds.

For 1c builds, as they have wider margins, occasional overfulls
should look better than wide inter-word spaces.

Get rid of the warning by removing the twocolumn option and applying
those minor tweaks.

Signed-off-by: Akira Yokosawa <akiyks@gmail.com>