Enable 'save' option of silence package
Pass an option 'save' to the silence package and let it save
filtered-out warnings in $basename.sil.
Tweak messages displayed at the end of builds and make it easy
to know where those warnings are saved.

For sans serif builds, print excerpt of warnings which are not
filtered out by the silence package.

Also add $basename.sil to .gitignore and Makefile's clean list.

Signed-off-by: Akira Yokosawa <akiyks@gmail.com>