runfirstlatex.sh: Tweak fatal error reports
runfirstlatex.sh is expected to catch fatal build errors of
"LaTeX Error:" and "pdfTeX error:".

Add check of "LaTeX Error:" and extend the grep range for
"pdfTeX error".

While at it, make the check of exitcode be really as late as
possible.

This change improves a message when a missing package is detected.

Signed-off-by: Akira Yokosawa <akiyks@gmail.com>