Enable 'trueslanted' option for newtxtext >= v1.65
Currently, microtype emits the following messages in .log files:

    Package microtype Info: Character `AE ' is missing
    (microtype)             in font `T1/ntxtlf/m/sl/10'.
    (microtype)             Ignoring protrusion settings for this character.
    Package microtype Info: Character `OE ' is missing
    (microtype)             in font `T1/ntxtlf/m/sl/10'.
    (microtype)             Ignoring protrusion settings for this character.
    Package microtype Info: Character `156' is missing
    (microtype)             in font `T1/ntxtlf/m/sl/10'.
    (microtype)             Ignoring protrusion settings for this character.
    Package microtype Info: Character `textexclamdown ' is missing
    (microtype)             in font `T1/ntxtlf/m/sl/10'.
    (microtype)             Ignoring protrusion settings for this character.
    Package microtype Info: Character `textquestiondown ' is missing
    (microtype)             in font `T1/ntxtlf/m/sl/10'.
    (microtype)             Ignoring protrusion settings for this character.

This is due to the tweak of using txfonts' slanted shape for newtx's.

As a matter of fact, newtxtext has gained an option 'trueslated', which
is a better option with wider character coverage.

Enable the option when newtxtext >= 1.65 is detected.

The detection is done in autodate.sh and reflected in newtxversion.tex.

newtx v1.65 was released in 2021 (exact date not listed in revision
history).

Signed-off-by: Akira Yokosawa <akiyks@gmail.com>
---
newtxtext.sty version on Ubuntu/Debian LTS and Fedora (as of Aug 26, 2023

Debian buster:   1.531  2018/03/27
Ubuntu focal:    1.531  2018/03/27
Debian bullseye: 1.630  2020/09/19
Ubuntu jammy:    1.705  2022/01/11
Debian bookworm: 1.71   2022/02/28
Fedora 37:       1.630  2020/09/19
Fedora 38:       1.71   2022/02/28