memorder: Use 'Linux-kernel memory model' as long description of LKMM
Still, "Linux-kernel memory-consistency model" is the official
name of LKMM and the Acronym page lists it.

Add an indexsee entry in the Index page:

    Linux-kernel memory model (LKMM)  See Linux-kernel memroy-
          consistency model (LKMM)

Apply those updated acronym macros in the now early sections
in memorder.tex.

Signed-off-by: Akira Yokosawa <akiyks@gmail.com>
---
make punctcheck now emits a complaint.
Need to update patterns in punctcheck.pl.