Undefine \IXpl and \IXhpl
\IXpl takes a single argumment and adds a plural-s implicitly in the
 printed text.  For example,

    \IXpl{memory barrier}

results in

    memory barriers

in printed text, with an index entry for "memory barrier".

However, this is not obvious in .tex sources and even I who defined
\IXpl sometimes need a few tens of seconds to remember the implicit
plural-s.

The purpose of \IXpl is to color indexed terms in -ix builds including
plural-s.

Mostly the same effect can be obtained by a simpler code with explicit
plural-s:

    \IX{memory barrier}s

Update the definition of \IXplx so that "address dependencies"
can be marked as:

    \IXplx{address dependenc}{ies}{y}

and results in "address dependency" in index.

Signed-off-by: Akira Yokosawa <akiyks@gmail.com>