Makefile: Add rules to use rsvg-convert for SVG --> PDF conversion
Recently, Inkscape has had a number of stability issues in both of
its GUI-mode and CLI-mode.  We have experienced crashes of CLI-
mode Inkscape caused by a change in glib, which can been resolved
by a pending update in the GTK3 library.  Commit f307e19f2be0
("Prevent inkscape from interacting with desktop manager") was
an ugly hack for mitigating such crashes.

rsvg-convert is a light-weight CLI-only converter.

Up until Ubuntu 20.04 LTS, rsvg-convert's coverage of SVG features
was not good enough for some of the figures in perfbook drawn by
using Inkscape.

rsvg-convert has improved since, and version 2.52.5 of rsvg-convert,
which is available on Ubuntu 22.04 LTS, and later versions do fairly
good jobs in this regard.

So let's use rsvg-convert when it is available.

For Ubuntu and Debian, rsvg-convert belongs to the librsvg2-bin
package.  For Fedora, it is provided in the librsvg2-tools package.

Signed-off-by: Akira Yokosawa <akiyks@gmail.com>