Plot RCU-test-ratio on-the-fly
RCU-test-ratio.eps has been updated regularly in response to new
releases of the mainline Linux kernel.

If such updates were to happen rarely, having the plot as EPS in the
Git repo would make sense.  It makes more sense to have only the data
in the repo and to do the plot on-the-fly as far as the plot is
concerned.

There is another thing to consider.

Using "set term postscript" would require groff as well as gnuplot
for building perfbook.  By using "set term svg", groff can be omitted
from the requirement list.

Update Makefile to add recipes for the on-the-fly plotting in SVG and
also update FAQ-BUILD.txt and gitlab-ci.yml to add the requirement of
gnuplot.

Signed-off-by: Akira Yokosawa <akiyks@gmail.com>
Cc: Leonardo Bras <leobras.c@gmail.com>