CodeSamples: Move struct declaration from .c to .h
In -O0 builds, forward declararation of struct in .h is insufficient
if there is code that need actual declaration of the struct with
in the .h.

Move the declaration from .c to .h.

This is not an issus with most GCC and CLANG versions with -O3
optimization level, but GCC as of Ubuntu 22.04 has trouble with.

Signed-off-by: Akira Yokosawa <akiyks@gmail.com>