Stop using "fvextra"
Currently, base latex engines are seeing epoch-making transition
in the hope of being able to generate accessible PDFs [1].

[1]: https://latex3.github.io/tagging-project/ "LaTeX Tagging Project"

Unfortunately, it is not possible to update all those contributed
LaTeX packages as needed by changes in core LaTeX.

Most of frequently used LaTeX packages are under the rader of the
devs and have "first aid" patches distributed along with the latex
base changes when necessary.

Since "latex2e <2025-06-01 PL2>", chapter and section titles in 2c
build's headers don't behave as expected.

It turns out that "lineno" package, which is required by "fvextra",
needs adjustment to work with recent releases of LaTeX2e [2].

[2]: https://github.com/latex-lineno/lineno/issues/20
     "Compatibility with recent LaTeX output routine, in twocolumn mode"

Fortunately, there is only two instances of code snippets who depend
on "fvextra".  As a tentative workaround, tweak each of them to make it
fit in both of columnwidth of twocolumn and textwidth of ebook size
and tweak leftside margin of code snippets in twocolumn so that their
right edges won't be too close to the text in its neighboring column.

Signed-off-by: Akira Yokosawa <akiyks@gmail.com>
---
Change Listing 2.1 and 2.2 to lising* so that they are rendered
nearby.  It does not look nice but further tweaks can wait, I suppose.